==> 1_Alpha_Vantage_API.ipynb <==
import pandas as pd
# from alpha_vantage.timeseries import TimeSeries
import requests
import json 
import time
API_key = '44R0O0ZHLA8X3XJO'

symbols = ['AAPL','MSFT','AMZN' ,'TSLA']
           
           
# ,'FB', 'GOOGL']

# You can get the full list of S&P 500 ticker symbols from Wikipedia and replace the symbols list with the 500 companies.
# This hasn't been done due to the limited API request possibility.

# data = pd.read_html('https://en.wikipedia.org/wiki/List_of_S%26P_500_companies')  # data from Wikipedia
# symbol_list = data[0]
# symbol_list.head()

# symbols = symbol_list['Symbol'].to_list()

# Apple Inc. (AAPL)
# Microsoft Corp. (MSFT)
# Amazon.com Inc. (AMZN)
# 4. Facebook Inc. (FB)
# Alphabet Inc. Class A Shares (GOOGL)


data = []

for symbol in symbols:
    
    params = {
    'function' : 'TIME_SERIES_DAILY_ADJUSTED',
    'symbol' : symbol,
#     'outputsize' : #full', #full timeframe data
    'outputsize' : 'compact', #limit to 100 rows
    'datatype' : 'json',
    'apikey' : API_key
    }

    url = 'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' + symbol + '&apikey=44R0O0ZHLA8X3XJO'
    
#     time.sleep(13)  # pause request not to break the API / get blocked
    
    response = requests.get(url, params)
    data.append(pd.DataFrame(response.json()['Time Series (Daily)']))  
    # get each symbol's data to a list from the json file and transform it to a dataframe from the Time Series table

    
# data 

stock_data = pd.concat(data).T  # transpose 

stock_data = stock_data['1. open']  # filter to column

# stock_data.columns = symbols # apply symbols as column name

# stock_data

# stock_data[0:365] # limit to last 365 days


stock_data_new = pd.concat(data).T

stock_data_new.reset_index(inplace = True)

stock_data_new = stock_data_new.melt(id_vars = 'index', var_name = ['Company'], value_name = 'Price')

stock_data_new.head(501)

# def standardize(col):
#     """
#     This function takes a column from a dataframe 
#     and returns a standardized column by subtracting the column's mean
#     and dividing by the column's standard deviation.
#     """
    
#     # your code here
#     the_mean = np.mean(col) 
#     the_std = np.std(col)
    
#     return (col - the_mean)/the_std
        

def return_inv(col):
    
    return ((col-col.shift(-1))/col.shift(-1)) * 100 #formula for calculated stock return 


stock_data = stock_data.astype(float) # change data type to float


stock_data[['AAPL_returns', 'MSFT_returns', 'AMZN_returns', 'TSLA_returns']] = stock_data[['AAPL', 'MSFT', 'AMZN', 'TSLA']].apply(return_inv)
#adding new rows with the calculated return price fileds 

stock_data

# stock_data.to_csv('stock_data.csv')

stock_data_full = pd.concat(data).T  # transpose 

stock_data.reset_index(inplace = True) # reset indexing

stock_data = stock_data.melt(id_vars = 'index', var_name = ['Company'], value_name = 'Price') # all_in_one wide-to-long version magic

stock_data

stock_data.to_csv('stock_data_close.csv') # export data
