==> 2_manually_pulling_data_for_each_field.ipynb <==
import pandas as pd
# from alpha_vantage.timeseries import TimeSeries
import requests
import json 
import time
API_key = '44R0O0ZHLA8X3XJO'

symbols = ['AAPL','MSFT','AMZN' ,'TSLA']
data = []

for symbol in symbols:
    
    params = {
    'function' : 'TIME_SERIES_DAILY_ADJUSTED',
    'symbol' : symbol,
#     'outputsize' : #full', #full timeframe data
    'outputsize' : 'compact', #limit to 100 rows
    'datatype' : 'json',
    'apikey' : API_key
    }

    url = 'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' + symbol + '&apikey=44R0O0ZHLA8X3XJO'
    
#     time.sleep(13)  # pause request not to break the API / get blocked
    
    response = requests.get(url, params)
    data.append(pd.DataFrame(response.json()['Time Series (Daily)']))  
    # get each symbol's data to a list from the json file and transform it to a dataframe from the Time Series table

    
# data 

stock_data = pd.concat(data).T  # transpose 

stock_data = stock_data['1. open']  # filter to column

stock_data.columns = symbols

stock_data.reset_index(inplace = True)

stock_data = stock_data.melt(id_vars = 'index', var_name = ['Company'], value_name = 'Open Price')

stock_data

# from alpha_vantage.timeseries import TimeSeries
import requests
import json 
import time
API_key = '44R0O0ZHLA8X3XJO'

symbols = ['AAPL','MSFT','AMZN' ,'TSLA']
data = []

for symbol in symbols:
    
    params = {
    'function' : 'TIME_SERIES_DAILY_ADJUSTED',
    'symbol' : symbol,
#     'outputsize' : #full', #full timeframe data
    'outputsize' : 'compact', #limit to 100 rows
    'datatype' : 'json',
    'apikey' : API_key
    }

    url = 'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' + symbol + '&apikey=44R0O0ZHLA8X3XJO'
    
#     time.sleep(13)  # pause request not to break the API / get blocked
    
    response = requests.get(url, params)
    data.append(pd.DataFrame(response.json()['Time Series (Daily)']))  
    # get each symbol's data to a list from the json file and transform it to a dataframe from the Time Series table

    
# data 

stock_data = pd.concat(data).T  # transpose 

stock_data = stock_data['2. high']  # filter to column

stock_data.columns = symbols

stock_data.reset_index(inplace = True)

stock_data_high = stock_data.melt(id_vars = 'index', var_name = ['Company'], value_name = 'High Price')

stock_data_high

# from alpha_vantage.timeseries import TimeSeries
import requests
import json 
import time
API_key = '44R0O0ZHLA8X3XJO'

symbols = ['AAPL','MSFT','AMZN' ,'TSLA']
data = []

for symbol in symbols:
    
    params = {
    'function' : 'TIME_SERIES_DAILY_ADJUSTED',
    'symbol' : symbol,
#     'outputsize' : #full', #full timeframe data
    'outputsize' : 'compact', #limit to 100 rows
    'datatype' : 'json',
    'apikey' : API_key
    }

    url = 'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' + symbol + '&apikey=44R0O0ZHLA8X3XJO'
    
#     time.sleep(13)  # pause request not to break the API / get blocked
    
    response = requests.get(url, params)
    data.append(pd.DataFrame(response.json()['Time Series (Daily)']))  
    # get each symbol's data to a list from the json file and transform it to a dataframe from the Time Series table

    
# data 

stock_data = pd.concat(data).T  # transpose 

stock_data = stock_data['3. low']  # filter to column

stock_data.columns = symbols

stock_data.reset_index(inplace = True)

stock_data_low = stock_data.melt(id_vars = 'index', var_name = ['Company'], value_name = 'low Price')

stock_data_low

symbols = ['AAPL','MSFT','AMZN' ,'TSLA']
data = []

for symbol in symbols:
    
    params = {
    'function' : 'TIME_SERIES_DAILY_ADJUSTED',
    'symbol' : symbol,
#     'outputsize' : #full', #full timeframe data
    'outputsize' : 'compact', #limit to 100 rows
    'datatype' : 'json',
    'apikey' : API_key
    }

    url = 'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' + symbol + '&apikey=44R0O0ZHLA8X3XJO'
    
#     time.sleep(13)  # pause request not to break the API / get blocked
    
    response = requests.get(url, params)
    data.append(pd.DataFrame(response.json()['Time Series (Daily)']))  
    # get each symbol's data to a list from the json file and transform it to a dataframe from the Time Series table

    
# data 

stock_data = pd.concat(data).T  # transpose 

stock_data = stock_data['6. volume']  # filter to column

stock_data.columns = symbols

stock_data.reset_index(inplace = True)

stock_data_volume = stock_data.melt(id_vars = 'index', var_name = ['Company'], value_name = 'Volume')

stock_data_volume

stock_data_volume.to_csv('stock_data_volume.csv')
stock_data_low.to_csv('stock_data_low.csv')
stock_data_high.to_csv('stock_data_high.csv')
stock_data.to_csv('stock_data_open.csv')


# Documents/GitHub/API_stockexchange/

symbols = ['AAPL','MSFT','AMZN' ,'TSLA']
data = []

for symbol in symbols:
    
    params = {
    'function' : 'TIME_SERIES_DAILY_ADJUSTED',
    'symbol' : symbol,
#     'outputsize' : #full', #full timeframe data
    'outputsize' : 'compact', #limit to 100 rows
    'datatype' : 'json',
    'apikey' : API_key
    }

    url = 'https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=' + symbol + '&apikey=44R0O0ZHLA8X3XJO'
    
#     time.sleep(13)  # pause request not to break the API / get blocked
    
    response = requests.get(url, params)
    data.append(pd.DataFrame(response.json()['Time Series (Daily)']))  
    # get each symbol's data to a list from the json file and transform it to a dataframe from the Time Series table

    
# data 

stock_data = pd.concat(data).T  # transpose 

stock_data = stock_data['5. adjusted close']  # filter to column

stock_data.columns = symbols

stock_data.reset_index(inplace = True)

stock_data_close = stock_data.melt(id_vars = 'index', var_name = ['Company'], value_name = 'Adj Close')

stock_data_close.to_csv('stock_data_close.csv')
