==> 3_Merged.ipynb <==
import pandas as pd

data_open = pd.read_csv('stock_data_open.csv')
data_close = pd.read_csv('stock_data_close.csv')
data_low = pd.read_csv('stock_data_low.csv')
data_high = pd.read_csv('stock_data_high.csv')
data_volume = pd.read_csv('stock_data_volume.csv')

data = pd.merge(data_open, data_close, left_on = ['index', 'Company'], right_on = ['index', 'Company']) #merging data

data = pd.merge(data, data_low, left_on = ['index', 'Company'], right_on = ['index', 'Company'])

data = pd.merge(data, data_high, left_on = ['index', 'Company'], right_on = ['index', 'Company'])
data = pd.merge(data, data_volume, left_on = ['index', 'Company'], right_on = ['index', 'Company'])

data = data[['index', 'Company' , 'Price' , 'Adj Close' , 'low Price' , 'High Price' , 'Volume']] #filtering

data

def return_inv(col):
    
    return ((col-col.shift(-1))/col.shift(-1)) * 100 #stock return calculation in %

data[['returns']] = data[['Adj Close']].apply(return_inv)

data.to_csv('final_data.csv')
