==> 4_Web_scraping_symbols_from_Wiki.ipynb <==
import pandas as pd

# https://medium.com/python-data/how-to-scrape-information-of-s-p-500-listed-companies-with-python-8205f895ee7a

data = pd.read_html('https://en.wikipedia.org/wiki/List_of_S%26P_500_companies')

table = data[0]
table.head()

# # no need to use it but in theory this would slice the bable and cut the column and row
# sliced_table = table[1:]  
# header = table.iloc[0]

# # update header names
# corrected_table = sliced_table.rename(columns=header) 
# corrected_table

# # exctract the tickers to a list
# tickers = corrected_table['Ticker symbol'].tolist()
# print (tickers)
